requirements.txt
# Core ML/DL
torch>=2.0.0
transformers>=4.35.0
datasets>=2.15.0
tokenizers>=0.15.0

# Cryptography
cryptography>=41.0.0
PyNaCl>=1.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Visualization & Evaluation
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
redis>=5.0.0
aiofiles>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipywidgets>=8.1.0
